requirements.txt
telethon
aiohttp
python-dotenv
cryptg